requirements.txt
streamlit
pdfplumber
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
openai
tiktoken